IDAxiBrillBH: Fix from Thomas Radke for floating point exception on the Alphas caused by exponention by a real number .... here's Thomas's description:
The include file generated by Mathematica contains an exponentiation
operation
with a REAL exponent.
My Fortran book says that

  y**x = exp(x * ln(y))

On O2K this intrinsic function checks the exponent against 0,
on Alphas it doesn't and you get a FP exception if y <= 0.

I just changed the type of the exponent to be INTEGER instead of REAL
which was probably also the intention of the programmer.


git-svn-id: https://svn.einsteintoolkit.org/cactus/EinsteinInitialData/IDAxiBrillBH/trunk@8 0a4070d5-58f5-498f-b6c0-2693e757fa0f