IDBrillData: Tidying so that the conformal solver isn't used to solve for the conformal factor, in fact the conformal factor is now only set at the end of the thorn.
Now if the conformal metric is used, psi will be set to the conformal factor
which was actually solved for. This means that right now that you can't use IDBrillData with conformal derivatives, although it would presumably be
easy enough to numerically calculate them, or to return 1 and 0's instead
just for this case.


git-svn-id: https://svn.einsteintoolkit.org/cactus/EinsteinInitialData/IDBrillData/trunk@76 a678b1cf-93e1-4b43-a69d-d43939e66649