Exact: Fix a bug spotted by Peter Diener:
The param.ccl and documentation.tex define the Kerr spin parameter to be
(dimensionless) J/m^2.  But (before this commit) the code used it with
(dimension = mass) J/m semantics.  This meant that the spin was set
wrong whenever the BH mass was different from 1.0

This change changes the code to properly match the documentation.

git-svn-id: https://svn.einsteintoolkit.org/cactus/EinsteinInitialData/Exact/trunk@126 e296648e-0e4f-0410-bd07-d597d9acff87