Exact: add an important thing Exact is lacking right now:
many people:
Every model should have .ge. 1 test case!  Right now most of the                models have no test cases at all, and this makes it way too easy for
bugs to creep in and go undetected for a long time.

For models satisfying the Einstein eqns, it would be nice to also have
test cases which compute the Hamiltonian constraint (and maybe the
momentum constraints too) and check that it (they) are suitably small.
        (The down side of that is that is that those test
        cases would then need at least ADMConstraints
        to be compiled into your Cactus binary, and that
        in turn requires ADMMacros.)
This would help guard against typos, wrong formulas, and other
"consistent, but wrong" cases.
~

git-svn-id: https://svn.einsteintoolkit.org/cactus/EinsteinInitialData/Exact/trunk@130 e296648e-0e4f-0410-bd07-d597d9acff87