Exact: add the "-->" part to the TODO entry:
Peter Diener:
There's no documented way to generate an entire exact spacetime
(not just a t=0 slice).  The "slice_evolver.tex" stuff may be able
to do this, but it looks complicated. :(  There ought to be a
simple way of doing this!  (And it ought to be documented simply, too!)

--> Peter has come up with a small patch to the param.ccl and schedule.ccl
    files which does this.  I (JT) need to document this and commit it
    to CVS.

git-svn-id: https://svn.einsteintoolkit.org/cactus/EinsteinInitialData/Exact/trunk@132 e296648e-0e4f-0410-bd07-d597d9acff87