Exact: this is now done:
Most of this thorn is C or Fortran 77.  There are only a few Fortran 90
files, and they're only used when computing initial data on an arbitrary
slice (as opposed to a t=constant slice).  It would be nice if there
were a clean/simple way to optionally disable that code/functionality
so this thorn could be compiled on systems without a Fortran 90 compiler.

git-svn-id: https://svn.einsteintoolkit.org/cactus/EinsteinInitialData/Exact/trunk@144 e296648e-0e4f-0410-bd07-d597d9acff87