Exact: add a note:
This thorn is mostly written in Fortran 77, with a few files in C
and Fortran 90.  At present Fortran 90 is only used for the "arbitrary slice"
option (see the thorn guide for details).  If this option isn't needed,
then by changing a single line in  src/include/Exact.inc  (see the comments
there for details) you can #ifdef-out the Fortran 90 code, allowing the
rest of this thorn to be compiled on a system with no Fortran 90 compiler
available.

(this was already noted in the thorn guide)

git-svn-id: https://svn.einsteintoolkit.org/cactus/EinsteinInitialData/Exact/trunk@167 e296648e-0e4f-0410-bd07-d597d9acff87