IDAnalyticBH: Segfault with kerr initial data using IDAnalyticBH
Yosef's patch:

If initial_data=kerr and metric_type=physical, IDAnalyticBH
writes to psi[x-z] and psi[x-z][x-z] (it tries to set these
GFs to zero) even if no storage is allocated. The problem
is with the code:

 IDAnalyticBH_zero_CCTK_REAL_array(npoints, psix);
 etc.., which is only executed when the metric is
 physical.

I was able to use the Kerr data with this patch, but just
deleting the offending lines should be sufficient.

git-svn-id: https://svn.einsteintoolkit.org/cactus/EinsteinInitialData/IDAnalyticBH/trunk@178 6a3ddf76-46e1-4315-99d9-bc56cac1ef84