ReadInterpolate: fix off by one error in ghost zone computation
I need to include a point whenever it is inside of the outermost ghost
point, not at least at the first real point.

Beyond that some guards against roundoff were introduced, the logic to
fill points made more robust and a check if all points were populated
was introduced.

On the down side we now run in level mode which means that without
further optimazations that HDF5 files are read once for each target
refinement level. This is needed to look like an ID thorn so
prim2coninitial sees our data.