TOVSolver: respect ADMBase::initial_shift parameter
Before this commit TOVSolver extended ADMBase::initial_shift to allow a "tov"
keyword value.  However it did not actually check anywhere that this value was
set and instead set the shift to zero unconditionally if TOV_Clear_Initial_Data
and not TOV_Use_Old_Initial_Data (line 546 of tov.c).

It should instead have checked the value of this parameter and set the shift to
zero if the parameter says "tov" independent of TOV_Clear_Initial_Data etc.

git-svn-id: https://svn.einsteintoolkit.org/cactus/EinsteinInitialData/TOVSolver/trunk@148 1bdb13ef-5d69-4035-bb54-08abeb3aa7f1